requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
bs4==0.0.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
distro==1.9.0
fonttools==4.58.4
frozenlist==1.7.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
Markdown==3.8.1
multidict==6.4.4
numpy==2.3.0
openai==1.86.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydyf==0.11.0
PyJWT==2.10.1
pyparsing==3.2.3
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
twilio==9.6.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
yarl==1.20.1
zopfli==0.2.3.post1
